Add Subscription model.
Why:

In order for a user to get updates on projects, searches, and more, they
need to be able to subscribe to these entities to watch them.

This change addresses the need by:

* Create a Subscription model to track 'watching' relationships between
users and other entities. A polymorphic relationship ensures we can
watch more than one model.

* Validating that the record is somehow related to developments, in
order to get a full changeset from related models.

* Future background job will use 'last_checked_at' timestamp to
determine how far back in the development history to look for changes.