Generate PDF reports from searches.
Why:

In order to let users download summary reports of their search results,
we needed a PDF summary that included a URL to the report / search.

This change addresses the need by:

* Create a template to show a map, summary statistics, and search
metadata.

* Auto-saving searches, with a default title, whenever a report is
generated from them. (Still to do: add an editable title to
the Search API resource.)

When there's some spare time after the first release, it would be a good
idea to come back to this and revise code around the iteration,
building, and accessing of report data.