Gracefully handle empty search params.
Why:

Sometimes, the client will return params like '?filter[commsf]=&',
meaning that the filter on :commsf gets triggered, but with an empty
string as a value.

This change addresses the need by:

* Returning a range of -Infinity to +Infinity for any empty values. That
is, nil, empty arrays, empty strings, etc. will return an infinite
range, so as not to exclude values just because we're missing a filter
value.

Note: I was originally going to return nil when parsing empty/false-y
values, but realized that would require nil checks and conditionals.
Returning Infinities was more true to the solution, and required no such
checks.