Check if subscribables need update.
Why:

We need to check whether the entities to which a user subscribes are in
need of update.

This change addresses the need by:

Testing and implementing whether users' subscriptions need to be
updated, by delegating most of the logic to the model itself.

This is done by checking whether the subscribed-to entity has been
updated - whatever that means for that entity - since the last time the
user had their subscriptions checked.