Add maps and simple street view.
Why:

In building out the development show page, we want to show maps and
photos.

This change addresses the need by:

* Adding a small leaflet map to sidebar, and placing a marker on the
development.

* Update the history feed to use the actual gravatars of users.

* Adding a Google Street View image.

NOTES:

* Street View is bound to be erroneous, as it is fixed in a certain
direction. It's nice to have a placeholder image, though. We can address
the requirements of properly implementing this later.

* It appears that when a user creates an organization, they don't
automatically become an admin & have a membership to that
organization.