Establish daily digest.
Why:

The digest mailer needed a lot of preparatory work to get it going, and
we're now at the point where we can start building it out.

This change addresses the need by:

* Starting a mail template, which still needs a lot of work.
* Altering the way objects are presented to make for a cleaner template.
* Ensure that the user's date of last subscription check is non-null.

The mailer needs a lot more work, especially around limiting how many
edits and changes are shown, the style (should look like the feed), and
links to the site.