Make development team unique; update sample data
Why:

An organization should not be able to be associated with the same role
on a development twice, though it should be able to play different roles
on the same development.

In order to have robust sample data, we needed to add
development teams and subscriptions.

This change addresses the need by:

* Testing and ensuring that :role is unique to the combination of
organization and development.
* Adding subscriptions and development teams to the sample data.
* Changes random model lookups in sample data from .all.sample to
ordering randomly. This prevents the need to instantiate all the records
before sampling, and should significantly increase the speed of the
sample data task.