Display organization members.
Why:

The organization page didn't display people clearly. Also, when
displaying a member-less organization (no longer likely to happen with
that last commit), there was no placeholder or call to action.

This change addresses the need by:

* Changes people display to list.
* Adds UNTESTED #primary_organization to user presenter.
* Removing the current_user check before showing the membership button,
and replaces it with authentication before joining. This will hopefully
encourage more people to sign up.