Rewrite API routes with Versionist, version module
Why:

The Versionist gem provides assistance in versioning API URLs, but the
syntax clutters the routes file to an obnoxious level. As we add
versions, there will be a lot of code duplication and even more noise.
We want simpler syntax.

Due to Rails magic, the api_version helper is available in routes.rb,
but for now seems like too much work just to get it in a helper class.

This change addresses the need by:

* Add an APIVersion class that returns the parameters for the
api_version helper method in routes.rb.

* Test APIVersion class.

* Remove APIConstraints, which we were using previously to version the
API, before introducing Versionist.