Add Gravatars to Organizations
Why:

We needed to start using organization avatars, and wanted to remove the
placeholder avatar thumbnails we were using.

We're also trying to delay requiring assets (and therefore S3 and
Paperclip) for as long as possible, because that always complicates
things. It would be nice to do that as one of the last steps before the
first release.

This change addresses the need by:

* Migrate gravatar_email and hashed_email columns on Organization.

* Use the Gravatar email or the primary email to generate a Gravatar
hash before saving the organization.

* Set gravatar_url, logo methods on Organization (not presenter). It
would be best in presenter, but there will need to be some wrapping
before we get there.

* Add links to development team, so that clicking on a development team
member routes to that organization, and adjust the layout to present it
cleanly.

* Add a valid Gravatar email to an organization in the sample data.