Add search information to JavaScript variables
Why:

In order to search from the Ember app, the app needs to know if there is
a logged-in user, what their API key is, and what the current min-max
values of searches are.

This change addresses the need by:

* Adding a namespace to the search page's JavaScript, and assigning it
to a window variable for presumably easier access.

* Creating methods on the Development class to calculate the bounds.

NOTE: If we ever add caching, it might be smart to add the min-max to
the cache, and recalculate on the background after every development
save, to avoid calculating this for every page load.