Add subscription helper methods to User
Why:

It's been a long time since I've gone through the process of adding an
AJAX-y follow/unfollow button. While we'll end up using the
APIController for the AJAX button, I think if we go through the process
of a server-side refresh, then the AJAX button, we'll end up with a
better-written end product.

This change addresses the need by:

* Adding subscribe, unsubscribe, and subscribed? helper methods to the
user class.