Start adding button to subscribe to developments.
Why:

We need to let users subscribe to developments. There should be a simple
AJAXy button to do this.

This change addresses the need by:

* Adding a labeled button to list subscribers.

* Adding a test to start testing this subscribe button.

* Altering button CSS to not have a gap between buttons and
left-pointing labels. (There's a margin on button that keeps them
spaced, but also pushes away labels.)

* Renaming a feature test from noun- to verb-oriented.

* Changing the password hashing cost in test fixtures from an
(incorrect) integer to the hashing library's constant MIN_COST.

* Remove integration tests.