Begin to revive development/show
Why:

The development show page has been empty of attributes since the first
time we tried to over-engineer attribute management. As we get closer to
implementing an editing interface, it would be good to start to present
the development data as a basis for the design.

NOTE: There are A LOT of broken tests, most of them integration.

This change addresses the need by:

* Adds helpers to the development presenter that categorize attributes.
There is an easy refactor apparent, to be done in a later commit.

* Adds conditionals to the template that encourage users to contribute
information.

* Makes contributor avatars into links to profiles

* Fixes a couple of N+1 queries