Add data to API representation, sample
Why:

The API was only returning a small subset of attributes for
developments. Additionally, we saw a lot of sample addresses with two
cities, like

  5320 Tristin Lights, Ivoryshire MA 88209-4170, Boston MA

We traced this down to the sample address being populated with city, but
we've since moved that to a Place relationship.

This change addresses the need by:

* Add attributes to API resource

* Fixing the sample data task to only have the street address, not city
etc.