Make the digest email succinct & presentable
Why:

We want to send an email -- weekly or daily -- that doesn't overwhelm
users, and presents plenty of hooks to jump into the site.

This change addresses the need by:

* Adding a summary presenter to limit changes to the first 3, with a
link to see more. (It doesn't lead to more, but rather to the
development.)

* Remove the 'See all activity' link on developments/show because we
don't yet have a view/controller for that.

* Add a migration and enumerization for an email frequency preference.
We don't have a form yet, but that will be in an upcoming commit.

* Customize the digest email for the user. Instead of "Daily" hardcoded,
use the frequency the user sets. Instead of a hardcoded Date.yesterday,
do it the last time the user had their subscriptions checked.

Note that this introduces a bug: if a user sets their email preference
to :never for a while, then sets it to something besides :never, they'll
receive a TON of changes and even generating the email might break the
server. We'll address this in an upcoming commit.

* Clean up the DigestPresenter, making it more Rubyish.

* Lay out the email with a container and lists.

* Make the digest presentation a partial, and make the user info in the
feed a partial.

* Change the list language -- it's showing how many of the entities the
user subscribes to were updated, not how many entities the user
subscribes to.

Note: this probably breaks a lot of tests, esp. CSV, since we add a new
attribute.