Prevent catastrophic mail bug.
Why:

If a user were to set their mail preference to :never for a really long
time, then come back to email, their last_checked_subscription attribute
would be a long time, and they'd get a ton of changes.

This change addresses the need by:

* Preveting the bug by setting a reasonable last_checked date whenever
the user changes from a :never to non-never email frequency preference.