Add street view fields and cache image
Why:

We are adding street view to give the page more visual appeal and
improve the presentation of information. The edit interface will allow
the user to set a street view location, pitch, and heading, which we'll
store to query the Google Street View Image API on the show page.

Instead of hitting the Google Street View Image API every time an image
is displayed, we want to cache it to prevent rate-limiting. However, we
also want to

This change addresses the need by:

* Adding Street View fields to the Development model, including pitch,
heading, and a  separate lat/lon so the user can refine street view but
keep the development's true location and dependent parcel reference.

* Adding methods to cache the street view image. NOTE: We should add a
mini preview image as well.

* Testing with webmock.

* Fixing up serializer tests that failed because street view was
introduced.

NOTE It would be easy -- and might make a lot of architectural sense --
to make StreetView its own model. It has enough attributes, and could
belong_to :development. Eventually, we could keep a history of the
street view of a development as it changes over time, and have a cool
historical retrospective.