Present street view
Why:

This was a confusing spate of work, and the architecture isn't fully
understood. I was going to add small street view images as well, but how
to do that isn't clear to me yet. I think more testing will reveal this,
but for now we just need street view up and running.

This change addresses the need by:

* Encoding the image for display

* Changing the way development/index is displayed

* Starts to test width and height options, but doesn't fully implement,
that is, doesn't get to the view layer