Add Membership, Organization authorizers
Why:

Since the object model around Developments and Edits is a bit more
complex, let's start simple: editing an organization, and joining it.

We don't have an admin role on the Membership class, which we'll need to
add. For now, we default to an admin that is simply the creator of the
organization. Organization roles will come later.

We want users to be able to request to create and join organizations. We
want organization admins to be able to change membership statuses, and
to edit organizations.

On architecture:

I considered having a can_join? method on the authorizer. While this
might be a good convenience method at some point, to think about the
verb 'join' in an object-oriented way, it means to create a membership
that links the user to the organization. So, user.can_join?(org) should
be asked, user.can_create?(Membership, on: organization).

This change addresses the need by:

* Adding Membership and Organization authorizers.