Handle unauthorized actions with a flash
Why:

Instead of displaying a jarring 'Unauthorized' error, we want to direct
people back to the page they were on, with a flash.

NOTE: We'll want to come back and test this.

NOTE: We also need to decide how to handle unauthorized errors in the
API.

This change addresses the need by:

* Adding the flash-based authorizer to the ApplicationController.