Prepare member, org relationships for authorizers
Why:

We are about to start on the Development and Edit Authorizers, core
permissions models for the entire application. To do that, we need to
square away how people and organizations are related to developments and
edits.

This change addresses the need by:

* Adding a 'municipal' boolean to organizations, because municipal orgs
will have more power in moderation.

* Adding a 'declined' option to organization memberships, so org admins
can prevent people from joining.

* Clarifying how people are related to developments -- that is, calling
any member of an organization that is a DevelopmentTeamMember a
moderator. This may change as permissions change in the future, but for
now, it's true.