Refactor validations, scopes on relations.
Why:

After just a few weeks away, with fresh eyes, I found a number of
low-hanging fruit for refactoring. A lot of these interfaces -- methods
like #incentive_programs -- seemed unnecessary and verbose to me at the
time, but I didn't have the clarity of vision to see why. I was so
focused on making the code read like English, I temporarily forgot that
code is not English -- it is code.

We had been setting the same validation code on the latitude and
longitude coordinates, which was causing visual noise and bloat.

This change addresses the need by:

This commit replaces the over-English syntax for better relationships
and scopes. Instead of #incentive_programs, we define a scope
Program.incentive. If we need a development's incentive programs, we can
call @development.programs.incentive. Instead of changes_since, we set a
scope Edit.history and Edit.since, which we can chain. So, instead of
changes_since, we have history.since.

We set the latitude and longitude validation options as variables, and
put those in the validations.