Boolean scopes look up 'nil's and 'false's
Why:

In the search endpoint, when we specify 'false' for a given boolean
attribute (e.g. 'stalled'), we want to return everything that's not
true, be it a nil or a false in the database.

Also, returning to the Periscope class after a few weeks, it took
significant mental effort to piece together what was going on. Clearly
needed documentation.

This change addresses the need by:

* Refactoring the Periscope module for additional clarity. Adding
significant documentation to the Periscope module.

* THEN, adding a condition so that false OR nil values would result in a
query returning records that are both false AND nil.

* Changing the Guardfile to automatically run tests in lib/ and
subdirectories, and moving tests of code in lib/extensions into
test/lib/extensions.

* Meanwhile, refactoring queries/options passed to multiple
API::V1::SearchController tests.