Edit tagline, look up Place, show WalkScore
Why:

This is a lot for a commit.

Generating a tagline from attributes is over-engineering at the moment.
For now, we make that an editable field so users can summarize it
themselves.

We decided that the best way to make the association to place is to use
the map location to look up neighborhood and/or municipality, instead of
having them select it through a dropdown or something. Also, just
parsing the Google geocoded address for text would be potentially
error-prone.

This change addresses the need by:

* Make tagline editable through API, and validate length on tagline and
description.

* When a development is saved, look up the underlying neighborhood or
municipality. We implement a case statement to custom-sort, so that
neighborhoods are listed before municipalities, thus getting the finest
geography first.

* If no municipality is found, we send Airbrake a notice, but we don't
raise an error.

* Display WalkScore on development/show.

* Test routes, in an attempt to get the Ember search and edit app
working.

* Delegate #municipality and #neighborhood to Place, instead of
implementing all that lookup logic in Development.