Access WalkScore more easily
Why:

WalkScore is kind of clunky, and may deserve its own serializer. We want
to know if it's empty -- that is, if it has a score -- and if so, to
access that score without needing a hash accessor.

This change addresses the need by:

* Allowing WalkScore to be instantiated from a hash, pulled from the
database -- without hitting the API again.

* Wrapping the Development#walkscore reader in a WalkScore object
instantiated from a hash.

* Autoloading lib, including WalkScore