Debug out-of-date partial
Why:

Needed to add the partial only when the development is out of date.

Also, there's a difference in the behavior of flash.now and flash --
when it goes through a request (so, 'flash' instead of 'flash.now'), it
converts the hash keys to strings but doesn't make access indifferent.

This change addresses the need by:

* Add a conditional on show view to add flash (.now) if about to show an
out of date development.

* Make the hash indifferent before accessing.