Fix failing tests.
Why:

A lot of tests had fallen out of line, or were catching genuine errors.

This change addresses the need by:

* Update serializer tests to not include walk score or parcel info in
the serializer, and to coerce BigDecimals to floats for the right
CSV output.

NOTE: Realized that, instead of these giant blobs of unwieldy test, we
could just load in a CSV fixture file, which would be easier to see and
edit to update tests. Should do that in next commits.

Might also be able to get away with using fixtures instead of the giant
'create' blobs at the top of serializer tests.