Fix approval of conflicts, pending edit display
Why:

We couldn't visit /edits/pending, because we were redirected by the
ember app route. Because the route was specified as '/edit', it was
catching '/edit' and anything that came after.

Declining conflicts was no problem -- but accepting was, especially when
there was a conflict.

The error message was out of sync as well. Some methods (like
\#unignored_conflict?) was unnecessarily private. Pending edits could
also be applied -- which is a contradiction in terms. Also, it was hard
to see the conflict or diff on an edit.

This change addresses the need by:

* Automatically overriding conflicts. If moderators can see the button,
they should be able to override them. I think I put too much custom
permisssions logic in, when validations would have been a better way to
think about this.

* Fixing the flash error that appears during 'unexpected errors'.

* Ensuring applied: false is specific for pending edits.

* Added a commented-out validation that should be completed when there's
time to refactor some.

* Adding helper methods on edits/changes to see diffs and conflicts more
easily.

* Add a presenter to show the conflicts as descriptive text.

* Rewrite Ember route so that it doesn't look for /edits/pending, but
stops matching at /edit.