Deflect development/update to proposed edit
Why:

With the Ember interface, we need to update developments through the
API, but the API wants to update the resource directly. We need to
deflect the call, creating a proposed edit based on the changes.

Tests were failing because we took attribute aliases out of the
resource, intending to redefine them in the model's FieldAliases module,
but never actually redefined them.

This change addresses the need by:

* Propose an edit instead of updating the development, and render
nothing successfully. This isn't as good as redirecting to another API
action like edits/show, but it's hard to figuring out how to override
the controller actions.

* Redefining attribute aliases.

* Changing tests to catch up to fixtures.