Import organizations
Why:

As we begin migrating data, I'm starting with organizations. Sam has
been working on a spreadsheet with developer info, and this commit
starts to see it added in.

This change addresses the need by:

* Add a task that imports organizations from CSV and creates
appropriate memberships.

* Add 'phone' and 'address' attributes to organization, though these are
presently neither shown nor edited on the site. Good to have in our data
if we need it later, though.

* Add the CSV of organizations, in its current state. This will be
updated as we add more developments.

* Remove the website validation. Imported records aren't going to have
this.

* I noticed that if many organizations didn't have a gravatar email or
regular contact email, they would have the same identicon-style Gravatar
-- hashing an empty string returns the same hash each time, obviously.
I added 'name' to the list of attributes so that, no matter what, the
identicon for an organization will always be unique, even if it means
hashing the name and not the email.