Fix edit_development route
Why:

I had put a $ at the end of the /edit route to get it to
point to the Ember interface, and to prevent /edits/pending from
matching the Ember route.

I thought the $ would act as a regex token. I was wrong.

Instead, the edit_development path literally had a dollar sign in the
route, and while it loaded the Ember interface, Ember didn't know what
to do with the dollar sign.

This change addresses the need by:

* Replacing the dollar sign with a trailing slash.