Fix organization load
Why:

Organizations weren't loading full information because #strip! returns
nil instead of the string -- we needed to use #strip to return the
string.

Also, some organizations had emails with weirdly-encoded characters,
which appeared in the text editor as spaces but which were decidedly not
spaces, as #strip had no effect on them.

This change addresses the need by:

* Change #strip! to #strip

* Remove weird characters from CSV.

* Shorten rake task name.

* Add a scope on Membership, implement on organization.