Present developments; header, footer
Why:

After adding in real development data, I noticed some issues. One was
that, due to how quickly we were calling the Street View API, we were
getting rate limit errors in the form of images.

The wording of the out-of-date header was also a little clunky, and
errored out for developments with no history -- that is, imported
developments.

Searches were erroring when there were 'nil' values in the search query.

This change addresses the need by:

* Run a checksum on the street view image, and compare it to the
checksum of the 'Over Rate Limit' error image that Google sends.

* Rewriting the Out of Date header.

* Fix search query by rejecting nil values. The root cause of error was
too hard to trace.

* Refactor street view so I'm not getting in my own way.

* Remove 'watch', 'claim' buttons, highlight Edit in green.

* Style header so sign up / log in are on the right, and sign up is
bright.