Treat Place as a resource, configure pagination
Why:

Place was being rendered as 'neighborhood', which returned the whole
geometry. Now, Place is represented as a related resource.

Also, after adding in all the real development data, all of the
developments were loading in the search page. We need to paginate, if
only for the sake of the server.

This change addresses the need by:

* Moving 'neighborhood' to the 'place' relationship.

* Adding JSONAPI Resources paginator configuration.