Add first-pass permissions for moderators
Why:

The only people who should be able to moderate edits on a given development
are:

- Users who are members of organizations that are members of the
development's development team
- Users who are members of municipal organizations, whose municipality
is the same as the development's municipality

In this commit, and the few around it, we're doing a first-pass at
permissions. Eventually, we'll be adding in Authorizers to check views
and controllers much more robustly.

This change addresses the need by:

* Adding User#moderator_for? to determine whether they should be allowed
to moderate.

* In support of this, adding a Place relationship on Organizations,
requiring a municipal or to have a municipality. NOTE: This is not yet
added to the 'new organization' form.

* Adding view logic to prevent the Moderate button from appearing to
anyone other than moderators.

* Adding controller checks on flags and moderation to ensure the user
doing an action is a moderator.

* After discovering that a user who creates an org is given a pending,
not active membership, changing the after_create hook to create an
active membership.