Fix false/nil searching.
Why:

When we pass filter[attribute]=false to the server, we want 'false' to
really return both false values and null values.

This change addresses the need by:

* Adding the condition to the JSONAPI filter helpers.

This brings up an issue: that the way range filters and boolean filters
are applied in the API vs not in the API are slightly different, and
mostly duplicate code. This is an area that could use refactoring later
on.