Add text search to Organization
Why:

After adding pagination (since we have a large number of organizations),
when adding a development team member, we can only see the first 10 or
so organizations.

In order to prepare for the client to do autocomplete-style search, we
need a filter to find organizations with names that match the text.

This change addresses the need by:

* Add a simple ILIKE query on the name attribute of the Organization
model.

* Add a filter :search on the Organization resource.