Fix interception of edits to development
Why:

Previously, editing a development through the client would result in an
empty edit. This is because we weren't actually applying changes to a
development.

Also, invalid records wouldn't return validation errors.

This change addresses the need by:

* Copying in the #process_request logic of JSONAPI Resources 0.7.0 (VERY
different interface from master).

* Overriding the process_operations section (the 'else' part of the
'unless...else' block), to persist the development edit, and clear out
the changes before rendering the data back to the client.

* Also, some validations were flat-out wrong: .validate doesn't take the
:if option of the with_options block, so 'validate ..._coerced' was
running regardless of the condition. Strange.

Special thanks to @lgebhardt of cerebris/jsonapi-resources for pointing
me in the right direction.