Make pending memberships buttons.
Why:

Pending memberships were links. They should be nice buttons, and they
should be disabled when a user is an admin, so admins can't be
deactivated or re-promoted.

Also, found that name was being validated for people resetting their
password to recover their account. Some people don't have first and last
names -- this should only be validated for new records.

This change addresses the need by:

* Adding buttons and disabling where appropriate.

* Setting user name validations to :new_record?