Fix Walk Score, nearest transit, and layout
Why:

After a demo with Tim, we found that WalkScore, parking spaces, and
nearest transit station were missing.

We also decided that Street View was taking up too much page space, and
might confuse people if it was oriented incorrectly. Also, images that
say "no image for this area" weren't being caught when we asked if the
Street View is null.

This change addresses the need by:

* Fixed a bug with WalkScore, where passing an option to :hash would
cause it to not be updated.

* Added an MBTA transit station lookup. If there are no major transit
stations, it uses the nearest bus stop.

* Alter layout, moving Street View to the sidebar and adding context to
help the user understand it's not perfect.

* Add 'no images in this location' to the list of null street view
images, so it doesn't display on the show page.

* Added "Last Updated" to empty "Recent Changes" feeds.

* Gave tables footers for rows that represent totals.