Fix image in Twitter card.
Why:

Found https://cards-dev.twitter.com/validator, which showed that the
card wasn't returning the image properly. This is because we had some
`street_view.to_s` being called, not the URL to the image.

But there is no URL for the image, so we needed to create one.

This change addresses the need by:

* Adding a route /developments/:id/image that shows just the binary
image.

* Linking to the image route in the Twitter card.