Add geospatial point, spatial bounding box query.
Why:

In order to load developments as a user scans a map, we need to be able
to send the bounding box of the map to an API endpoint and load all of
the developments within the bounds.

* It is easiest to do this with a spatial query, so we add a geometry
attribute 'point'.

* We set the record to update the point if latitude or longitude
change.

* We add a filter :bbox that takes WGS84 lat-lon values west, south,
east, north and returns records within that box.

* We also add a "within" spatial query to the Development model, which
intersects points with an arbitrary geometry.