Add geocoding (address) search to searchables
Why:

The frontend Google geocoder/autocomplete wasn't working well with
Ember, so we decided to include the search in the backend, and serving
it out as a simple result.

This change addresses the need by:

* Creating a MapzenSearch service object, which looks up text using
Mapzen autocomplete.

* Adding the location to the top of the results,

NOTE: This may need significant tuning. We should also focus the results
dynamically when we have time to do so:
https://trello.com/c/uUXb47Ah/153-focus-the-search.