Rework exports
Why:

Previously, one had to save a search in order to export that search. We
want people to be able to export anything.

This change addresses the need by:

* Adding a developments/export route and controller action that take in
search params, and generate an export/report based on that -- no saving
needed.

* Makes a filename for the search out of the first 2 search criteria.

* The map referenced #results, which is incorrect since #developments is
the public method that limits the results to the first 100 - a limit
we're putting in until we have time to set up a background queue.
The method #results should really be private to the search/show
template.

* Change links to reference the new paths, instead of the saved search.

NOTE: Should have a default title for an unsaved search based on params,
e.g. 'Projected between 100 and 200 housing units' and the filename just
dasherizes the title.