Add image URL, Arrayify a nil
Add an image_url so that the thumbnails on the client-side search can
link to an image path.

When there's an error or no Mapzen API key, we get a nil response.
Arrayify that so it can't be nil, ever.

Include municipality in the neighborhood default_scope. It's basically
always referenced.