Remove validation on unit/sqft sums
Reality check: a lot of projects which are not yet "in construction" or
"completed" have partial information about the unit/sqft breakdowns,
which don't total the total number of units/sqft

Removes the validation on projects that aren't "in construction" or
"completed".

Refactors residential, commercial attribute names as constants on the
model. Why hadn't we done this before?