Move WalkScore JSON to string
Why:

We've been wanting to use .uniq on scopes, but cannot because Postgres
9.3 cannot handle DISTINCT on tables with JSON columns. There are
hacky workarounds which, soon, will produce memory bloat -- and
upgrading Postgres isn't a viable option at the moment.

This change addresses the need by:

* Migrating the JSON to a string of JSON.

Once this migration is done, the next commit should be to deserialize
the walkscore string as JSON when loading from the database.