Block abusive clients
Why:

Reading through the logs, we see a lot of scrape-style requests, and
some hacking attempts. We want to block these.

This change addresses the need by:

* Adding rack-attack middleware and a blocklist to prevent access.

NOTE: We should probably have some nofollow attributes to prevent
scraping certain pages, like flags/new, which appears a lot in the logs.