Refactor edit approval, decline service objects
Why:

Decided to namespace, and to provide more uniform logic. Hopefully, this
is more legible to readers, and the namespacing will clear the way for
more services.

This is also a warmup for heavier refactoring.

This change addresses the need by:

* Adding default conditional to check whether the service is callable.

* Defining and commenting where to add logic.

* Removing unnecessary bits -- like shaking an architectural model. A
lot simply didn't need to be there, or could be moved into the
superclass.