Refactor routes
Why:

Clarify what's happening in the routes, so it's simpler to understand.

This change addresses the need by:

* Renaming SubdomainConstraint to Constraint, because the only thing
that made it a subdomain constraint was the regex and a method inside.

* Refactor to be more flexible -- call the regex match on any part of
the request, from the routes file, by passing :host or some other
method.

* Change APIVersion to be a module, instead of a method, to indicate its
function and that it has an immediate return.