Development test refactoring with DI
Why:

The Development test tested too much. By using dependency injection, I
was able to simplify the tests and test external services independently.

The fakes used in the tests, are themselves a little brittle, but it's
better than making a ton of webmock calls everywhere -- now they only
happen where they need to.

We test external services elsewhere, and when we need to know the model
update causes a change, we just use the fake to make sure something gets
called.

This commit also adds the 40B, etc. "projects" attribute.