Remove conflict cruft from the code
Why:

Originally, we thought we'd need to protect merge conflicts from
occurring through the backend. In reality, when there's a conflict, we
just need to ensure that the view requires the moderator to take an
extra step to review the conflict.

This change addresses the need by:

* No longer requiring the :ignore_conflicts flag to be set when
approving an edit.

* Removing helper methods #conflict?, etc. from edits, and using
`.conflicts.any?` to check for conflicts