Fix #159 by removing validations
This removes the first and last name validations for users because users should not have to enter this information to have an account, and previous users that had not entered this information were getting errors when resetting their passwords. This commit fixes this problem by no longer validating first and last names. It also updates relevant tests and views to gracefully handle the lack of a first or last name from a user model.