Renders proper layout
A previews change removed the specialized Ember template layout and had it default to something else. It seems like it was some form of refactoring, so this adds the :new action back in. If we use Ember for the :edit action again, we need to make sure it renders the ‘search’ layout.