Fix #122 By Using Search API
The autocomplete API was not returning results from Mapzen with street addresses. This was happening for two reasons. The first is that the key we were giving it had commas, and commas throw exceptions, which is why we override the verify_key method to strip out commas. The second is we were using the autocomplete API which stops returning useful feature results when the strings get too specific. So we switched to the search API which does not have this issue.