Pick Most Confident Result from Mapzen
Mapzen’s search API returns ten results with varying confidence, but is not sorted by that. Although we were getting a useful answer from Mapzen, our own API was discarding it because it only looked at the first result and required the confidence to be more than 0.75. This commit fixes this by sorting the results from Mapzen by confidence and then returns a result if the confidence is greater than 0.6. This should provide the Ember client something to use in a wider variety of situations.